Extracted SwaggerResourcesProvider to an interface
Also created an default In-memory swagger resources provider that will be
available to be overridden by a @Primary bean implementation for providing
aggregation of services

closes #1189
