+++ springfox-swagger-common/src/main/java/springfox/documentation/swagger/web/SwaggerResourcesProvider.java
@@ -1,0 +1,26 @@
+/*
+ *
+ *  Copyright 2015 the original author or authors.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ *
+ *
+ */
+package springfox.documentation.swagger.web;
+
+import com.google.common.base.Supplier;
+
+import java.util.List;
+
+public interface SwaggerResourcesProvider extends Supplier<List<SwaggerResource>> {
+}
